requirements.txt
django
uvicorn
psycopg2
markdown
django-filter 
django-crispy-forms
crispy-bootstrap5
django-bulk-update
coverage
channels
Twisted[tls,http2]
hiredis
channels-redis
celery
django-celery-beat
django-environ
bleach
cryptography
pandas
numpy
alive-progress
requests
watchfiles